Sikandar_Bakht/sprint1/SprintOneProj/requirements.txt
aws-cdk.core>=1.135.0
aws-cdk.assertions>=1.135.0
aws-cdk.aws.lambda>=1.135.0
aws-cdk.core>=1.135.0
aws-cdk.aws_events_targets>=1.135.0
aws-cdk.aws_cloudwatch_actions>=1.135.0
aws-cdk.aws_dynamodb>=1.135.0